// interfaces/mytarjet/usersSegments-interface.ts
// Generated by https://quicktype.io

export interface UsersMyTarjetResultInterface {
    SDTTarjetsG: SDTTarjetsG[];
}

export interface SDTTarjetsG {
    IdUsuario:      string;
    IdTarjet:       string;
    Status:         string;
    NombreCompleto: string;
    UsuEncabezado:  string;
    UsuActividad:   string;
    UsuFondoF:      string;
    UsuImagenQR:    string;
    UsuUrl:         string;
    UsuActiva:      number;
    SegmentoId:     string;
    SegmentoDesc:   string;
    UsuTitulo:      string;
    UsuTituloDesc:  string;
    UsuToken:       string;
    ImgFoto:        string;
    Cargo:          string;
}